requirements.txt
datasets==2.14.5
huggingface-hub==0.16.4
hydra-core==1.3.1
numpy==1.23.5
scikit-learn==1.2.0
timm==0.5.4
torch==2.0.1+cu117
TorchCRF==1.1.0
tqdm==4.64.0
transformers==4.31.0